Add camp website link for past event: Auckland 2025
I'd been thinking previously that it'd be good to have links to the websites of past events, so I've added a link to that camp's website. I'm sure that site probably won't be around forever, so thought I'd use an archive.org link, but the latest version there is not yet up to date. We could switch these to archive links when they go down?